+++ crates/fluke-buffet/Cargo.toml
@@ -15,8 +15,8 @@ rust-version = "1.75.0"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [features]
-default = ["tokio-uring"]
-tokio-uring = ["dep:tokio-uring"]
+default = ["uring"]
+uring = ["dep:fluke-io-uring-async"]
 miri = []
 
 [dependencies]
@@ -24,7 +24,6 @@ bytemuck = { version = "1.15.0", features = ["extern_crate_std"] }
 eyre = "0.6.12"
 http = "1.1.0"
 io-uring = "0.6.3"
-fluke-io-uring-async = { path = "../fluke-io-uring-async", version = "0.1.0" }
 libc = "0.2.153"
 memchr = "2.7.1"
 memmap2 = { version = "0.9.4", default-features = false }
@@ -44,7 +43,7 @@ tracing = "0.1.40"
 nix = "0.28.0"
 
 [target.'cfg(target_os = "linux")'.dependencies]
-tokio-uring = { git = "https://github.com/tokio-rs/tokio-uring", rev = "a69d4bf57776a085a6516f4c022e2bf5d1814762", optional = true }
+fluke-io-uring-async = { path = "../fluke-io-uring-async", version = "0.1.0", optional = true }
 
 [dev-dependencies]
 color-eyre = "0.6.3"
