--- crates/fluke-buffet/src/compat.rs
@@ -1,1 +1,0 @@
-
