+++ crates/fluke-buffet/src/io/non_uring.rs
@@ -1,8 +1,4 @@
-use crate::{
-    buf::{IoBuf, IoBufMut},
-    io::{ReadOwned, WriteOwned},
-    BufResult, Piece,
-};
+use crate::{BufResult, IoBufMut, Piece, ReadOwned, WriteOwned};
 
 use tokio::io::{AsyncRead, AsyncWrite, AsyncWriteExt};
 
@@ -11,14 +7,8 @@ where
     T: AsyncRead + Unpin,
 {
     async fn read<B: IoBufMut>(&mut self, mut buf: B) -> BufResult<usize, B> {
-        let buf_slice =
-            unsafe { std::slice::from_raw_parts_mut(buf.stable_mut_ptr(), buf.bytes_total()) };
+        let buf_slice = unsafe { buf.slice_mut() };
         let res = tokio::io::AsyncReadExt::read(self, buf_slice).await;
-        if let Ok(n) = &res {
-            unsafe {
-                buf.set_init(*n);
-            }
-        }
         (res, buf)
     }
 }
@@ -28,8 +18,7 @@ where
     T: AsyncWrite + Unpin,
 {
     async fn write(&mut self, buf: Piece) -> BufResult<usize, Piece> {
-        let buf_slice = unsafe { std::slice::from_raw_parts(buf.stable_ptr(), buf.bytes_init()) };
-        let res = tokio::io::AsyncWriteExt::write(self, buf_slice).await;
+        let res = tokio::io::AsyncWriteExt::write(self, &buf[..]).await;
         (res, buf)
     }
