+++ crates/fluke/Cargo.toml
@@ -14,8 +14,8 @@ rust-version = "1.75"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [features]
-default = ["tokio-uring"]
-tokio-uring = ["fluke-buffet/tokio-uring"]
+default = ["uring"]
+uring = ["fluke-buffet/uring"]
 
 [dependencies]
 byteorder = "1.5.0"
