+++ crates/fluke/src/h1/body.rs
@@ -4,8 +4,7 @@ use tracing::debug;
 
 use crate::{util::read_and_parse, Body, BodyChunk, BodyErrorReason};
 use fluke_buffet::{
-    io::{ReadOwned, WriteOwned},
-    PieceCore, PieceList, RollMut,
+    PieceCore, PieceList, RollMut, {ReadOwned, WriteOwned},
 };
 
 /// An HTTP/1.1 body, either chunked or content-length.
